plainbox:unit:job: add support for pre/post command
This patch adds two new optional fields that are applicable to all
command-bearing jobs (so all except manual and qml). The pre and post
commands are, as expected, supposed to be started before and after
running the main command.

This is not specified yet but there's some extra interaction that may
need to happen for log files (as now command can have additional log
files) and we also need to address the issue of losing log files after
re-running a job.

The error codes for both commands are likely going to be discarded
though I suspect the pre-command could also fail the job if it fails.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>