plainbox:session:state: add _bumps_modified decorator
This decorator's purpose is to bump the 'modified_timestamp' to current time.
All methods that change the state of session in a way that's later reflected in
the checkpoint stored on the drive should be decorated with it. (Or to keep it
simple - evey method that changes the state object).

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>