plainbox:session:suspend: add datetime proxy mock
This patch introduces a proxy that eases the process of developing unit tests
that test  stuff calling datetime.datetime.utcnow(). datetime is a built-in so
the only way to mock it traditionally is to mock the whole object, but in the
tests to come some of the original datetime functionality needs to be preserved
(e.g. isoformat). Similar effect could be achieved without using padme with a
code like this: http://paste.ubuntu.com/13628916/

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>